+++ src/Either.php
@@ -165,4 +165,28 @@ final class Either
     {
         return $this->either->memoize();
     }
+
+    /**
+     * Switch the sides of the values, left becomes right and right left
+     *
+     * @return self<R, L>
+     */
+    public function flip(): self
+    {
+        return new self($this->either->flip());
+    }
+
+    /**
+     * @template A
+     * @template B
+     *
+     * @param callable(R): self<A, B> $right
+     * @param callable(L): self<A, B> $left
+     *
+     * @return self<A, B>
+     */
+    public function eitherWay(callable $right, callable $left): self
+    {
+        return $this->either->eitherWay($right, $left);
+    }
 }