+++ src/Either/Implementation.php
@@ -83,4 +83,20 @@ interface Implementation
      * @return Either<L, R>
      */
     public function memoize(): Either;
+
+    /**
+     * @return self<R, L>
+     */
+    public function flip(): self;
+
+    /**
+     * @template A
+     * @template B
+     *
+     * @param callable(R): Either<A, B> $right
+     * @param callable(L): Either<A, B> $left
+     *
+     * @return Either<A, B>
+     */
+    public function eitherWay(callable $right, callable $left): Either;
 }