+++ src/Either/Left.php
@@ -81,4 +81,15 @@ final class Left implements Implementation
     {
         return Either::left($this->value);
     }
+
+    public function flip(): Implementation
+    {
+        return new Right($this->value);
+    }
+
+    public function eitherWay(callable $right, callable $left): Either
+    {
+        /** @psalm-suppress ImpureFunctionCall */
+        return $left($this->value);
+    }
 }