+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 4.14.0 - 2023-04-29
+
+### Added
+
+- `Innmind\Immutable\Either::flip()`
+- `Innmind\Immutable\Maybe::toSequence()`
+- `Innmind\Immutable\Maybe::eitherWay()`
+- `Innmind\Immutable\Either::eitherWay()`
+
 ## 4.13.0 - 2023-04-10
 
 ### Added